requirements.txt
accelerate==1.1.1
aioboto3==13.2.0
aiobotocore==2.15.2
aiocache==0.12.3
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioitertools==0.12.0
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
anytree==2.12.1
asttokens==2.4.1
attrs==24.2.0
autograd==1.7.0
backoff==2.2.1
beartype==0.18.5
boto3==1.35.36
botocore==1.35.36
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
executing==2.1.0
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
gensim==4.3.3
graspologic==3.4.1
graspologic-native==1.2.1
h11==0.14.0
hnswlib==0.8.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
hyppo==0.5.1
idna==3.10
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.7.1
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
jsonpickle==4.0.0
kiwisolver==1.4.7
lightrag-hku==1.0.1
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
nano-vectordb==0.0.4.3
neo4j==5.26.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
ollama==0.3.3
openai==1.54.4
oracledb==2.5.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.0.0
POT==0.9.5
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.27.0
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
s3transfer==0.10.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
torch==2.2.2
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.2
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
urllib3==2.2.3
wcwidth==0.2.13
wrapt==1.16.0
xxhash==3.5.0
yarl==1.17.1